treewide: initial commit of OpenRoaming configuration scripts/packages
To ease deployment of the required configuration for OpenRoaming, and
lend ourselves to more reproducible builds. Drop in an initial commit
of the minimum required set of configuration, patches, and boilerplate
to stand up infrastructure to enable ANP functionality and/or client
functionality.

RADIUS and radsec components are segregated from general wireless
configuration as they may not be deployed onto the same device.
Though for our purposes of demonstration, they often are.

Include patches to
 - busybox - to enable naptr queries in nslookup;
 - radsecproxy - to update to 1.11.2
 - freeradius - to enable rfc4072 support and coa module enabled;
 - netifd - to support configuring a client with a credential block;

These patches are a bit awkward, and if demand allows, should be
upstreamed to relevant repositories.

Signed-off-by: Arien Judge <arien.judge@morsemicro.com>